==> code/01_clean_data.R <==

##----------------------- PARENTAL LOSS FROM DRUGS/FIREARMS --------------------
##
##  Code cleaning the data used to study the impact of drugs and firearms on
#   parental loss in the US, stratifying by race/ethnicity, and gender.
##  The impact will be assessed using the matrix kinship model
##  developed by Hal Caswell.
## 
## 
##  Author: Benjamin Schlüter
##  Date: May 2023
##------------------------------------------------------------------------------
##
##  Notes
## ------
##  MORTALITY:
##
##  - Same n_a_x for all ethnic groups
##
##
##  FERTILITY:
##
##  - Compute fx for male by shifting female fx to the right by the 
##    difference between the mean age at childbearing.
##      - Same shift for all ethnic groups
##
##  - Model without male fertility?
##
##  - Intrapolate fx with Hermitte cubic (avoid wiggles of cubic-spline).
##
##
##  Checks
## -------
##  (Single race) in brth.xlsx: dropped but what does it represents?
## 
##------------------------------------------------------------------------------

rm(list = ls())



## Load packages ---------------------------------------------------------------

## Install/load packages
packages <- c("tidyverse", "ggplot2", "here", "openxlsx", "readxl", "Matrix",
              "pracma", "pdftools")
for(p in packages){
        if(!require(p,character.only = TRUE)) install.packages(p)
        library(p,character.only = TRUE)
}



## Functions -------------------------------------------------------------------

source(here("code", "fct_intrapol_fx.R"))


## Load data -------------------------------------------------------------------

## Mortality data from CDC, generated by Matt
df.dth <- readRDS(here("data_private", 
                       ## Kept original name of file 
                       ## Data is not public and hence
                       ## not accessible on github.
                       ## Use link from Matt and place it in "data_private"
                       "state_year_age-sex-race_drug-opioid-mortality.RDS")) %>% 
        rename("age" = age_years) %>% 
        dplyr::select(!c(st_fips, division))

## Merge all states to work at national level
df.dth <- df.dth %>% 
        group_by(year, race_eth, sex, age) %>% 
        summarise_at(vars(pop:n_firearm), sum) %>% 
        mutate(abbrev = "US",
               name = "United States")

## Add "total" as an additional ethnic group
df.dth.total <- df.dth %>% 
        group_by(year, sex, age) %>% 
        summarise_at(vars(pop:n_firearm), sum) %>% 
        ungroup() %>% 
        mutate(race_eth = "total",
               abbrev = "US",
               name = "United States")
df.dth <- rbind(df.dth,
                df.dth.total)

## Fertility data from CDC

## Table has a difficult format to import in R
## Careful: last age interval is 10 year wide (45-54)

## Excel cols of interest
columns.xlsx <- c(1, 4, 5, 8:13)
age.gp.xlsx <- seq(10, 45, 5)
columns.names <- c("year", as.character(age.gp.xlsx))
## Rows for each ethnic group from 2000 to 2019
rows.total.xlsx <- 31:50 # All races
rows.nhw.xlsx <- 119:142 # White, not Hispanic or Latina
rows.nhb.xlsx <- 200:223 # Black or African American, not Hispanic or Latina
rows.aian.xlsx <- 262:285 # American Indian or Alaska Native
rows.api.xlsx <- 320:339 # Asian or Pacific Islander, not Hispanic or Latina
rows.h.xlsx <- 362:381 # Hispanic or Latina
## Same order as appearance in Excel file 
## and same labels as mortality data
ethnic.gp <- c("total", "white", "black", "aian", "api", "hispanic")
years.xlsx <- 2000:2019
n.years.xlsx <- length(years.xlsx)

fx <- read.xlsx(here("data_raw", 
                     "brth.xlsx"),
                rows = c(rows.total.xlsx,
                         rows.nhw.xlsx,
                         rows.nhb.xlsx,
                         rows.aian.xlsx,
                         rows.api.xlsx,
                         rows.h.xlsx),
                cols = columns.xlsx,
                colNames = F) %>% 
        ## Rename cols
        rename_at(paste0("X", 1:9), ~ columns.names) %>% 
        # Remove year concerning "single race"
        filter(!grepl("(single race)", year)) %>% 
        mutate(
                ## "*": rate based on <20 births not shown-> set to 0.
                across(all_of(columns.names), ~ sub("*", 0, .x, fixed = T)),
               ## Add ethnic group variable
               race_eth = rep(ethnic.gp, each = n.years.xlsx)
               ) %>% 
        ## Horizontal to vertical data frame
        pivot_longer(
                as.character(age.gp.xlsx), 
                     values_to = "fx", names_to = "age"
                ) %>% 
        ## Define some variables as numeric
        mutate(across(c(year, age, fx), ~ as.numeric(.))) 

## Fertility data from CDC stops in the year 2019
## Get fertility rates from latest CDC report
## Source: https://www.cdc.gov/nchs/data/nvsr/nvsr72/nvsr72-01.pdf
## Consulted on the 24/05/2023
## Done manually but would be better to extract table from PDF 
## and tidy (though it might involve substantial coding while currently
## I simply copied/pasted)
## Age groups
age.fx <- unique(fx$age)
n.age.fx <- length(age.fx)

fx.latest.total <- tibble(year = 2020,
                          age = age.fx,
                          fx = c(0.2, 15, 63.3, 90.9, 94.9, 51.3, 11.8, 0.9),
                          race_eth = "total")

fx.latest.nhw <- tibble(year = 2020,
                          age = age.fx,
                          fx = c(0.1, 10.1, 53.1, 88.2, 98.2, 49.7, 10.1, 0.7),
                          race_eth = "white")

fx.latest.nhb <- tibble(year = 2020,
                        age = age.fx,
                        fx = c(0.3, 23.7, 84.0, 93.4, 80.4, 47.1, 12.7, 1.1),
                        race_eth = "black")

fx.latest.aian <- tibble(year = 2020,
                        age = age.fx,
                        fx = c(0.4, 25.2, 82.2, 83.2, 68.8, 35.9, 8.1, 0.3),
                        race_eth = "aian")

fx.latest.h <- tibble(year = 2020,
                         age = age.fx,
                         fx = c(0.3, 23, 83.8, 105.8, 94.4, 53.4, 14.2, 1.0),
                         race_eth = "hispanic")
## Api not available

fx.latest <- rbind(fx.latest.total,
                   fx.latest.nhw,
                   fx.latest.nhb,
                   fx.latest.aian,
                   fx.latest.h) 
        
        

## Bind with CDC data
fx <- rbind(fx,
            fx.latest) %>% 
        # Fx expressed per 1000 women
        mutate(fx = fx/1000) 


## Male fertility by age, race and year is not available from CDC.
## Use Table 10 & 11 from UN Demographic Yearbooks to get male 
## and female fertility rates for available years: 
## 2008, 2012, 2013, 2015 and 2019. 
## In what follows, we compute mean age at childbearing (MAC) for
## both sexes.  MAC will be used to generate male fx for each race.

## Data has a weird format to import and is not harmonized over the years.
## In addition, some births have an unknown father age so they had to be
## redistributed.
## -> Significant cleaning
## Cleaning is done first for female and then for male. One loop could do both
## sexes at once but for readability of code, leaved it as it is for now.
years.dyb <- as.character(c(2008, 2012, 2013, 2015, 2019))
columns.names <- c("age.gp", "birth")

## Female

## Elements used for cleaning in loop below 
## Rows corresponding to the US in Table 10 over the years found in 
## the UN Demographic Yearbook
t10.rows <- list(605:613,556:564,571:579,668:676,658:667)
names(t10.rows) <- years.dyb
age.gp.f <- c(0, seq(15, 55, 5)) ## Assume reproduction stops at 55 yo for female
mid.age.gp.f <- c(7.5, seq(17.5, 52.5, 5))

## Cleaning loop over the years for female
## and compute mean age at childbearing (MAC)
mac.f.dyb <- sapply(years.dyb, function(y) {
        
        ## Rows for US in Excel file in a given year
        rows.temp <- unlist(t10.rows[y])
        
        ## Load and tidy dyb data (ie harmonize age group)
        t10 <- read.xlsx(here("data_raw", 
                              paste0("t10_", y, ".xlsx")),
                         rows = rows.temp,
                         cols = 1:2,
                         colNames = F) %>% 
                ## Rename cols
                rename_at(paste0("X", 1:2), ~ columns.names) %>% 
                ## Harmonize age gp
                mutate(age.gp = sub("\\ .*", "", age.gp),
                       age.gp = as.numeric(age.gp),
                       age = cut(age.gp,
                                 breaks = age.gp.f,
                                 labels = head(age.gp.f, -1),
                                 right = F)) %>%
                group_by(age) %>%
                summarise(birth = sum(birth))

        ## Population counts at reproductive ages
        pop.repro <- df.dth %>%
                filter(race_eth == "total",
                       ## Assume female reproduction stops at 55 yo
                       age < max(age.gp.f),
                       sex == "female",
                       year == as.numeric(y)) %>%
                mutate(age = cut(age,
                                 breaks = age.gp.f,
                                 labels = head(age.gp.f, -1),
                                 right = F)) %>%
                group_by(year, sex, age) %>%
                summarise(pop = sum(pop)) %>%
                ungroup()

        ## Join the two and compute fertility rates
        fx <- t10 %>%
                left_join(pop.repro,
                          by = "age") %>%
                mutate(fx = birth/pop)

        ## Compute mean age at childbearing
        mac <- fx %>%
                mutate(mid.age = mid.age.gp.f) %>%
                group_by(sex, year) %>%
                summarise(MAC = sum(fx*mid.age)/sum(fx)) %>%
                pull(MAC)
        
        
        return(mac)
},
simplify = T, USE.NAMES = TRUE)

## Male

## Elements used for cleaning in loop below 
## Rows corresponding to the US in Table 11 over the years found in 
## the UN Demographic Yearbook
t11.rows <- list(283:292,275:284,275:284,218:227,239:248)
names(t11.rows) <- years.dyb
age.gp.m <- c(0, seq(20, 55, 5), 65) ## Assume reproduction stops at 65 yo for male
mid.age.gp.m <- c(10, seq(22.5, 52.5, 5), 60)

## Cleaning loop over the years for male
## and compute mean age at childbearing (MAC)
mac.m.dyb <- sapply(years.dyb, function(y) {
        
        ## Rows for US in Excel file in a given year
        rows.temp <- unlist(t11.rows[y])
        
        ## Load and tidy dyb data (ie harmonize age group & redistribute unknown)
        t11 <- read.xlsx(here("data_raw", 
                              paste0("t11_", y, ".xlsx")),
                         rows = rows.temp,
                         cols = 1:2,
                         colNames = F) %>% 
                ## Rename cols
                rename_at(paste0("X", 1:2), ~ columns.names) #%>% 
        
        ## Redistribute unknown in accordance with 
        ## the distribution of births by age of father 
        unknwn.birth <- t11 %>% 
                filter(age.gp == "Unknown - Inconnu") %>% 
                pull(birth)
        
        dist.birth <- t11 %>% 
                filter(age.gp != "Unknown - Inconnu") %>% 
                group_by() %>% 
                mutate(prop = birth/sum(birth)) %>% 
                pull(prop)
        
        redist.unknwn.birth <- unknwn.birth * dist.birth
        
        t11 <- t11 %>% 
                filter(age.gp != "Unknown - Inconnu") %>% 
                ## Redistribute unknown births
                mutate(birth = birth + redist.unknwn.birth) %>% 
                ## Harmonize age gp
                mutate(age.gp = sub("\\ .*", "", age.gp),
                       age.gp = as.numeric(age.gp),
                       age = cut(age.gp,
                                 breaks = age.gp.m,
                                 labels = head(age.gp.m, -1),
                                 right = F)) %>%
                group_by(age) %>%
                summarise(birth = sum(birth))

        ## Population counts at reproductive ages
        pop.repro <- df.dth %>%
                filter(race_eth == "total",
                       ## Assume female reproduction stops at 65 yo
                       age < max(age.gp.m),
                       sex == "male",
                       year == as.numeric(y)) %>%
                mutate(age = cut(age,
                                 breaks = age.gp.m,
                                 labels = head(age.gp.m, -1),
                                 right = F)) %>%
                group_by(year, sex, age) %>%
                summarise(pop = sum(pop)) %>%
                ungroup()

        ## Join the two and compute fertility rates
        fx <- t11 %>%
                left_join(pop.repro,
                          by = "age") %>%
                mutate(fx = birth/pop)

        ## Compute mean age at childbearing
        mac <- fx %>%
                mutate(mid.age = mid.age.gp.m) %>%
                group_by(sex, year) %>%
                summarise(MAC = sum(fx*mid.age)/sum(fx)) %>%
                pull(MAC)
        
        
        return(mac)
},
simplify = T, USE.NAMES = TRUE)

## Birth counts data come from two CDC National Vital Statistics Reports
## Years 1999:2015 come from National Vital Statistics Reports Vol 66, Nber 1
## Table 5 on p.24 
## Years 2016:2020 come from National Vital Statistics Reports Vol 72, Nber 1
## Table 1 on p.12 

## Years 1999-2015

txt <- pdf_text(here("data_raw", "t5_birth_2015_1989.pdf"))

## Table 5 is on page 24
table <- txt[24]
# cat(table5)

## Split the text into a one raw text matrix
text_matrix <- table %>% str_split("\\n", simplify = TRUE)
## Locate the start and end of the table
## ! make sure only one occurence or specify which one it is !
table_start <- text_matrix %>% 
        str_which("2015. . . . . . . . . . . . . . . . . . . . . . . . . . . . .")
table_end <- text_matrix %>% 
        str_which("611,269")

## Extract table text
table_raw <- text_matrix[1, table_start:table_end] %>%
        ## Remove period/dots next to dates
        str_replace_all(.,'[\\.,]','') %>% 
        ## Replace space with "|"
        str_replace_all("\\s{2,}", "|")
## Names of cols from PDF
column.names <- c("year", "total",
                  "hispanic", "mexican", "puertorican", "cuban", "csamerican", "other",
                  "nh", "white", "black")
## Create text connection so that the text can be read back with read.csv()
df.brth.2015 <- table_raw %>% 
        textConnection() %>% 
        read.csv(sep = "|", 
                 header = FALSE,
                 col.names = column.names) %>%
        as_tibble() %>% 
        ## Only look at years where we have mortality
        ## from opioids and firearms
        filter(year %in% 1999:2020) %>% 
        ## Only keep Hispanic, Black, and White
        dplyr::select(year, hispanic, white, black, total) %>% 
        ## Define variables as numeric
        mutate(across(colnames(.), ~ as.numeric(.))) %>% 
        pivot_longer(hispanic:total, names_to = "race_eth", values_to = "births")

## Years 2016-2020

txt <- pdf_text(here("data_raw", "t1_birth_2021_2016.pdf"))
## Table 1 is on p.12
table <- txt[12]

## Split the text into a one raw text matrix
text_matrix <- table %>% str_split("\\n", simplify = TRUE)
## Locate the start and end of the table
table_start <- text_matrix %>% 
        str_which("All races and origins")
table_end <- text_matrix %>% 
        str_which("70.6")

## Extract table text
table_raw <- text_matrix[1, table_start:table_end] %>%
        ## Remove period/dots next to dates
        str_replace_all(.,'[\\.,]','') %>% 
        ## Replace space with "|"
        str_replace_all("\\s{2,}", "|") 
## Remove lines without data
table_raw <- table_raw[-c(1, 14, 15, 22, 29, 36, 43, 50)]
## Some rows have "|" before the year, remove it
table_raw[13:42] <- sub("^.", "", table_raw[13:42])

## Names of cols from PDF
column.names <- c("year", "births", "birthrate", "fertilityrate")
## Years extracted in this pdf
years.brth <- 2016:2020
## Create text connection so that the text can be read back with read.csv()
df.brth.2020 <- table_raw %>% 
        textConnection() %>% 
        read.csv(sep = "|", 
                 col.names = column.names,
                 header = FALSE) %>%
        as_tibble() %>% 
        ## Only look at years in 2016-2020
        filter(year %in% as.character(years.brth)) %>% 
        ## Only keep births
        dplyr::select(year, births) %>% 
        ## Define ethnic group
        ## !! Same order as in source table !!
        mutate(race_eth = rep(c("total", "white", "black", "aian", "a", "pi", "hispanic"), 
                              each = length(years.brth))) %>% 
        ## Only keep ethnic groups of interest
        filter(race_eth %in% c("total", "white", "black", "hispanic"))

## Rbind the two tidy data sets
df.brth <- rbind(df.brth.2015,
                 df.brth.2020) %>% 
        arrange(year)




## Key dimensions for code ---------------------------------------------------------

## Analysis focus on years 2000:2020
years <- unique(df.dth$year)
n.years <- length(years)

## Ages considered
ages <- unique(df.dth$age)
n.ages <- length(ages)

## Races considered
races <- unique(df.dth$race_eth)
n.races <- length(races)

## Radix for life tables
radix <- 100000



## Life tables computation -----------------------------------------------------

## From a mortality perspective,
## quantities needed for kin dynamics are px all causes
## and qx by causes as they allow to create the U and M matrices.

lt.US <- df.dth %>%
        ## Assume same ax for all races/ethnicities
        group_by(year, race_eth, sex) %>% 
        mutate(mx = n_deaths/pop,
               ax = case_when(age == 0 ~ 0.07 + 1.7*mx[1],
                              age == max(ages) ~ Inf,
                              TRUE ~ 0.5),
               n = case_when(age == max(ages) ~ Inf,
                             TRUE ~ 1),
               qx = (n * mx)/(1 + (n - ax) * mx),
               qx = ifelse(age == max(ages), 1, qx),
               px = 1 - qx,
               lx = cumprod(c(1, head(px, -1)))*radix,
               dx = lx - lead(lx, 1),
               dx = ifelse(age == max(ages), lx, dx),
               Lx = lead(lx)*n + ax*dx,
               # Last age gp is NA due to lead()
               Lx = ifelse(age == max(ages), dx/mx, Lx),
               ex = rev(cumsum(rev(Lx)))/lx) %>%
        ungroup()



## Fertility computation -------------------------------------------------------

## Generate male fx for each ethnic group by shifting
## female fx according to the difference in MAC 
## over the years. 

## Create a df containing mac
df.diff.mac <- tibble(year = as.numeric(years.dyb),
                      female = mac.f.dyb,
                      male = mac.m.dyb) %>% 
        ## Compute difference in mac
        mutate(diff.mac = male - female) 

## Linear intra/extrapolation of diff.mac
lm.diff.mac <- lm(diff.mac ~ year, data = df.diff.mac)

## Predict for every year
df.fitted.diff.mac <- tibble(year = years) %>% 
        mutate(fit = predict(lm.diff.mac, newdata = .)) %>% 
        ## Join true values
        left_join(df.diff.mac %>% 
                          dplyr::select(!c(female, male)),
                  by = "year")

## Check: predicted diff.mac

# ggplot(df.fitted.diff.mac) +
#         geom_point(aes(x = year, y = diff.mac)) +
#         geom_line(aes(x = year, y = fit, group = 1)) +
#         theme_minimal()

## Interpolation of female fx by strata
fx.int.f <- fx %>% 
        ## Mid age class
        mutate(age.m = case_when(age == 45 ~ age + 5, # Last age group=45-54 yo
                                 TRUE ~ age + 2.5)) %>% 
        arrange(race_eth, year, age) %>% 
        group_by(race_eth, year) %>% 
        ## Own defined function
        group_modify(~ intra_fx(.x, sex = "female")) 

## Check: interpolation

# ggplot() +
#         facet_grid(race_eth ~ year) +
#         geom_point(data = fx %>% 
#                            mutate(age.m = case_when(age == 45 ~ age + 5, # Last age group=45-54 yo
#                                                     TRUE ~ age + 2.5)),
#                    aes(x = age.m, y = fx)) +
#         geom_line(data = fx.int.f, aes(x = age, y = fx, group = 1))

## Construct male fx by shifting female fx according to the fitted
## difference in mean age at childbearing 
fx.int.m <- fx %>% 
        ## Assume the same shift for all ethnic groups
        left_join(df.fitted.diff.mac,
                  by = "year") %>% 
        ## Mid age class
        mutate(age.m = case_when(age == 45 ~ age + 5, # Last age group=45-54 yo
                                 TRUE ~ age + 2.5),
               ## Shift fertility schedule according to fitted mac
               age.m = age.m + fit) %>% 
        ## Intrapolate 
        arrange(race_eth, year, age) %>% 
        group_by(race_eth, year) %>% 
        ## Own defined function
        group_modify(~ intra_fx(.x, sex = "male")) 

## Rbind female and created male fx
fx.int <- rbind(fx.int.f,
                fx.int.m)




## Save tidy data --------------------------------------------------------------

## Matt data can't be shared
saveRDS(lt.US,
        here("data_private", "lt_US.rda"))
saveRDS(fx.int,
        here("data", "fx_US.rda"))
saveRDS(df.fitted.diff.mac,
        here("data", "mac_US.rda"))
saveRDS(df.brth,
        here("data", "birth_US.rda"))
